Major rewrite
Reordered the guide to fit better the average new players need and added some extra direction on things noobs can do to get better at the game. Removed some weird part about things flying and killing you (?) and added two sections on the "tutorial" (how to put on the oxigen mask) and basic directions on how to use the radio. Added some clues to the mentor help chat that is about to be implemented.